todo-app-backend/requirement.txt
flask
flask_cors
python-dotenv
flask_sqlalchemy
sqlalchemy
psycopg2-binary